Move text area related code to /browser/
