+++ src/vs/editor/browser/controller/keyboardHandler.ts
@@ -8,8 +8,8 @@ import * as browser from 'vs/base/browser/browser';
 import * as dom from 'vs/base/browser/dom';
 import { IKeyboardEvent } from 'vs/base/browser/keyboardEvent';
 import { GlobalScreenReaderNVDA } from 'vs/editor/common/config/commonEditorConfig';
-import { TextAreaHandler } from 'vs/editor/common/controller/textAreaHandler';
-import { TextAreaStrategy } from 'vs/editor/common/controller/textAreaState';
+import { TextAreaHandler } from 'vs/editor/browser/controller/textAreaHandler';
+import { TextAreaStrategy } from 'vs/editor/browser/controller/textAreaState';
 import { Range } from 'vs/editor/common/core/range';
 import { ViewEventHandler } from 'vs/editor/common/viewModel/viewEventHandler';
 import { Configuration } from 'vs/editor/browser/config/configuration';
