+++ src/vs/editor/browser/controller/textAreaState.ts
