+++ src/vs/editor/browser/view/viewImpl.ts
@@ -20,7 +20,7 @@ import * as editorBrowser from 'vs/editor/browser/editorBrowser';
 import { ViewController, ExecCoreEditorCommandFunc } from 'vs/editor/browser/view/viewController';
 import { ViewEventDispatcher } from 'vs/editor/common/view/viewEventDispatcher';
 import { ContentViewOverlays, MarginViewOverlays } from 'vs/editor/browser/view/viewOverlays';
-import { LayoutProvider } from 'vs/editor/common/viewLayout/viewLayout';
+import { ViewLayout } from 'vs/editor/common/viewLayout/viewLayout';
 import { ViewContentWidgets } from 'vs/editor/browser/viewParts/contentWidgets/contentWidgets';
 import { CurrentLineHighlightOverlay } from 'vs/editor/browser/viewParts/currentLineHighlight/currentLineHighlight';
 import { CurrentLineMarginHighlightOverlay } from 'vs/editor/browser/viewParts/currentLineMarginHighlight/currentLineMarginHighlight';
@@ -66,7 +66,7 @@ export class View extends ViewEventHandler {
 
 	private eventDispatcher: ViewEventDispatcher;
 
-	private layoutProvider: LayoutProvider;
+	private layoutProvider: ViewLayout;
 	private _scrollbar: EditorScrollbar;
 	public _context: ViewContext;
 
@@ -121,7 +121,7 @@ export class View extends ViewEventHandler {
 		// - scrolling (i.e. viewport / full size) & co.
 		// - whitespaces (a.k.a. view zones) management & co.
 		// - line heights updating & co.
-		this.layoutProvider = new LayoutProvider(configuration, model.getLineCount(), this.eventDispatcher);
+		this.layoutProvider = new ViewLayout(configuration, model.getLineCount(), this.eventDispatcher);
 
 		// The view context is passed on to most classes (basically to reduce param. counts in ctors)
 		this._context = new ViewContext(configuration, model, this.eventDispatcher);
