+++ src/vs/editor/test/browser/controller/textAreaState.test.ts
@@ -5,11 +5,97 @@
 'use strict';
 
 import * as assert from 'assert';
-import { IENarratorTextAreaState, ISimpleModel, TextAreaState } from 'vs/editor/common/controller/textAreaState';
+import { IENarratorTextAreaState, ISimpleModel, TextAreaState, IClipboardEvent, ICompositionEvent, IKeyboardEventWrapper, ITextAreaWrapper } from 'vs/editor/browser/controller/textAreaState';
 import { Position } from 'vs/editor/common/core/position';
 import { Range } from 'vs/editor/common/core/range';
 import { EndOfLinePreference } from 'vs/editor/common/editorCommon';
-import { MockTextAreaWrapper } from 'vs/editor/test/common/mocks/mockTextAreaWrapper';
+import Event, { Emitter } from 'vs/base/common/event';
+import { Disposable } from 'vs/base/common/lifecycle';
+
+export class MockTextAreaWrapper extends Disposable implements ITextAreaWrapper {
+
+	private _onKeyDown = this._register(new Emitter<IKeyboardEventWrapper>());
+	public onKeyDown: Event<IKeyboardEventWrapper> = this._onKeyDown.event;
+
+	private _onKeyUp = this._register(new Emitter<IKeyboardEventWrapper>());
+	public onKeyUp: Event<IKeyboardEventWrapper> = this._onKeyUp.event;
+
+	private _onKeyPress = this._register(new Emitter<IKeyboardEventWrapper>());
+	public onKeyPress: Event<IKeyboardEventWrapper> = this._onKeyPress.event;
+
+	private _onCompositionStart = this._register(new Emitter<ICompositionEvent>());
+	public onCompositionStart: Event<ICompositionEvent> = this._onCompositionStart.event;
+
+	private _onCompositionUpdate = this._register(new Emitter<ICompositionEvent>());
+	public onCompositionUpdate: Event<ICompositionEvent> = this._onCompositionUpdate.event;
+
+	private _onCompositionEnd = this._register(new Emitter<ICompositionEvent>());
+	public onCompositionEnd: Event<ICompositionEvent> = this._onCompositionEnd.event;
+
+	private _onInput = this._register(new Emitter<void>());
+	public onInput: Event<void> = this._onInput.event;
+
+	private _onCut = this._register(new Emitter<IClipboardEvent>());
+	public onCut: Event<IClipboardEvent> = this._onCut.event;
+
+	private _onCopy = this._register(new Emitter<IClipboardEvent>());
+	public onCopy: Event<IClipboardEvent> = this._onCopy.event;
+
+	private _onPaste = this._register(new Emitter<IClipboardEvent>());
+	public onPaste: Event<IClipboardEvent> = this._onPaste.event;
+
+	public _value: string;
+	public _selectionStart: number;
+	public _selectionEnd: number;
+	public _isInOverwriteMode: boolean;
+
+	constructor() {
+		super();
+		this._value = '';
+		this._selectionStart = 0;
+		this._selectionEnd = 0;
+		this._isInOverwriteMode = false;
+	}
+
+	public getValue(): string {
+		return this._value;
+	}
+
+	public setValue(reason: string, value: string): void {
+		this._value = value;
+		this._selectionStart = this._value.length;
+		this._selectionEnd = this._value.length;
+	}
+
+	public getSelectionStart(): number {
+		return this._selectionStart;
+	}
+
+	public getSelectionEnd(): number {
+		return this._selectionEnd;
+	}
+
+	public setSelectionRange(selectionStart: number, selectionEnd: number): void {
+		if (selectionStart < 0) {
+			selectionStart = 0;
+		}
+		if (selectionStart > this._value.length) {
+			selectionStart = this._value.length;
+		}
+		if (selectionEnd < 0) {
+			selectionEnd = 0;
+		}
+		if (selectionEnd > this._value.length) {
+			selectionEnd = this._value.length;
+		}
+		this._selectionStart = selectionStart;
+		this._selectionEnd = selectionEnd;
+	}
+
+	public isInOverwriteMode(): boolean {
+		return this._isInOverwriteMode;
+	}
+}
 
 suite('TextAreaState', () => {
